@@ -174,124 +174,6 @@ module Refinery
       end
     end
 
-    context 'custom slugs' do
-      let(:custom_page_slug) { 'custom-page-slug' }
-      let(:custom_child_slug) { 'custom-child-slug' }
-      let(:custom_route) { '/products/my-product' }
-      let(:page_with_custom_slug) {
-        subject.class.new(:title => page_title, :custom_slug => custom_page_slug)
-      }
-      let(:child_with_custom_slug) {
-        page.children.new(:title => child_title, :custom_slug => custom_child_slug)
-      }
-      let(:page_with_custom_route) {
-        subject.class.new(:title => page_title, :custom_slug => custom_route)
-      }
-
-      after(:each) do
-        Refinery::I18n.stub(:current_frontend_locale).and_return(I18n.default_frontend_locale)
-        Refinery::I18n.stub(:current_locale).and_return(I18n.default_locale)
-      end
-
-      it 'returns its path with custom slug' do
-        page_with_custom_slug.save
-        page_with_custom_slug.url[:id].should be_nil
-        page_with_custom_slug.url[:path].should == [custom_page_slug]
-      end
-
-      it 'allows a custom route when slug scoping is off' do
-        turn_off_slug_scoping
-        page_with_custom_route.save
-        page_with_custom_route.url[:id].should be_nil
-        page_with_custom_route.url[:path].should == [custom_route]
-        turn_on_slug_scoping
-      end
-
-      it 'allows slashes in custom routes but slugs everything in between' do
-        turn_off_slug_scoping
-        page_needing_a_slugging = subject.class.new(:title => page_title, :custom_slug => 'products/category/sub category/my product is cool!')
-        page_needing_a_slugging.save
-        page_needing_a_slugging.url[:id].should be_nil
-        page_needing_a_slugging.url[:path].should == ['products/category/sub-category/my-product-is-cool']
-        turn_on_slug_scoping
-      end
-
-      it 'returns its path underneath its parent with custom urls' do
-        child_with_custom_slug.save
-        page.save
-
-        child_with_custom_slug.url[:id].should be_nil
-        child_with_custom_slug.url[:path].should == [page.url[:path].first, custom_child_slug]
-      end
-
-      it 'does not return a path underneath its parent when scoping is off' do
-        turn_off_slug_scoping
-        child_with_custom_slug.save
-        page.save
-
-        child_with_custom_slug.url[:id].should be_nil
-        child_with_custom_slug.url[:path].should == [custom_child_slug]
-        turn_on_slug_scoping
-      end
-
-      it "doesn't allow slashes in slug" do
-        page_with_slashes_in_slug = subject.class.new(:title => page_title, :custom_slug => '/products/category')
-        page_with_slashes_in_slug.save
-        page_with_slashes_in_slug.url[:path].should == ['productscategory']
-      end
-
-      it "allow slashes in slug when slug scoping is off" do
-        turn_off_slug_scoping
-        page_with_slashes_in_slug = subject.class.new(:title => page_title, :custom_slug => 'products/category/subcategory')
-        page_with_slashes_in_slug.save
-        page_with_slashes_in_slug.url[:path].should == ['products/category/subcategory']
-        turn_on_slug_scoping
-      end
-
-      it "strips leading and trailing slashes in slug when slug scoping is off" do
-        turn_off_slug_scoping
-        page_with_slashes_in_slug = subject.class.new(:title => page_title, :custom_slug => '/products/category/subcategory/')
-        page_with_slashes_in_slug.save
-        page_with_slashes_in_slug.url[:path].should == ['products/category/subcategory']
-        turn_on_slug_scoping
-      end
-
-      it 'returns its path with custom slug when using different locale' do
-        Refinery::I18n.stub(:current_frontend_locale).and_return(:ru)
-        Refinery::I18n.stub(:current_locale).and_return(:ru)
-        page_with_custom_slug.custom_slug = "#{custom_page_slug}-ru"
-        page_with_custom_slug.save
-        page_with_custom_slug.reload
-
-        page_with_custom_slug.url[:id].should be_nil
-        page_with_custom_slug.url[:path].should == ["#{custom_page_slug}-ru"]
-      end
-
-      it 'returns path underneath its parent with custom urls when using different locale' do
-        Refinery::I18n.stub(:current_frontend_locale).and_return(:ru)
-        Refinery::I18n.stub(:current_locale).and_return(:ru)
-        child_with_custom_slug.custom_slug = "#{custom_child_slug}-ru"
-        child_with_custom_slug.save
-        child_with_custom_slug.reload
-
-        child_with_custom_slug.url[:id].should be_nil
-        child_with_custom_slug.url[:path].should == [page.url[:path].first, "#{custom_child_slug}-ru"]
-      end
-
-      context "given a page with a custom_slug exists" do
-        before do
-          FactoryGirl.create(:page, :custom_slug => custom_page_slug)
-        end
-
-        it "fails validation when a new record uses that custom_slug" do
-          new_page = Page.new :custom_slug => custom_page_slug
-          new_page.valid?
-
-          new_page.errors[:custom_slug].should_not be_empty
-        end
-      end
-    end
-
     context 'content sections (page parts)' do
       before do
         page.parts.new(:title => 'body', :content => "I'm the first page part for this page.", :position => 0)
@@ -339,8 +221,8 @@ module Refinery
       it 'reposition correctly' do
         page.save
 
-        page.parts.first.update_attributes :position => 6
-        page.parts.last.update_attributes :position => 4
+        page.parts.first.update_columns position: 6
+        page.parts.last.update_columns position: 4
 
         page.parts.first.position.should == 6
         page.parts.last.position.should == 4
