@@ -1,6 +1,15 @@
 Refinery::Core::Engine.routes.draw do
   begin
     require 'devise'
+
+    # Override Devise's other routes for convenience methods.
+    devise_scope :refinery_user do
+      get "/#{Refinery::Core.backend_route}/login", :to => "sessions#new", :as => :login
+      get "/#{Refinery::Core.backend_route}/logout", :to => "sessions#destroy", :as => :logout
+      get "/#{Refinery::Core.backend_route}/users/register" => 'users#new', :as => :new_signup
+      post "/#{Refinery::Core.backend_route}/users/register" => 'users#create', :as => :signup
+    end
+
     devise_for :refinery_user,
                :class_name => 'Refinery::User',
                :path => "#{Refinery::Core.backend_route}/users",
@@ -9,14 +18,6 @@ Refinery::Core::Engine.routes.draw do
                :path_names => { :sign_out => 'logout',
                                 :sign_in => 'login',
                                 :sign_up => 'register' }
-
-    # Override Devise's other routes for convenience methods.
-    devise_scope :refinery_user do
-      get "/#{Refinery::Core.backend_route}/login", :to => "sessions#new", :as => :login
-      get "/#{Refinery::Core.backend_route}/logout", :to => "sessions#destroy", :as => :logout
-      get "/#{Refinery::Core.backend_route}/users/register" => 'users#new', :as => :signup
-      post "/#{Refinery::Core.backend_route}/users/register" => 'users#create', :as => :signup
-    end
   rescue RuntimeError => exc
     if exc.message =~ /ORM/
       # We don't want to complain on a fresh installation.
