@@ -142,11 +142,13 @@ module Refinery
       options = {:plural => false, :admin => true}.merge options
 
       klass = klass.constantize if klass.respond_to?(:constantize)
-      active_name = ActiveModel::Name.new klass, (Refinery if klass.parents.include?(Refinery))
+      active_name = ::ActiveModel::Name.new(
+        klass, (Refinery if klass.parents.include?(Refinery))
+      )
 
       if options[:admin]
         # Most of the time this gets rid of 'refinery'
-        parts = active_name.underscore.split('/').reject{|name|
+        parts = active_name.to_s.underscore.split('/').reject{|name|
           active_name.singular_route_key.exclude?(name)
         }
 
