@@ -9,12 +9,12 @@ module Refinery
 
       def thumbnail_urls(image)
         thumbnail_urls = {
-          :"data-original" => asset_paths.compute_public_path(image.url, ''),
-          :"data-grid" => asset_paths.compute_public_path(image.thumbnail(:geometry => '135x135#c').url, '')
+          :"data-original" => image_path(image.url),
+          :"data-grid" => image_path(image.thumbnail(:geometry => '135x135#c').url)
         }
 
         Refinery::Images.user_image_sizes.sort_by{|key,geometry| geometry}.each do |size, pixels|
-          thumbnail_urls[:"data-#{size.to_s.parameterize}"] = asset_paths.compute_public_path(image.thumbnail(:geometry => pixels).url, '')
+          thumbnail_urls[:"data-#{size.to_s.parameterize}"] = image_path(image.thumbnail(:geometry => pixels).url)
         end
 
         thumbnail_urls
