@@ -44,10 +44,17 @@ module Refinery
         ##
         # Injects Dragonfly::Middleware for Refinery::Images into the stack
         def attach!(app)
-          if ::Rails.application.config.action_controller.perform_caching
-            app.config.middleware.insert_after 'Rack::Cache', 'Dragonfly::Middleware', :refinery_images
+          if defined?(::Rack::Cache)
+            unless app.config.action_controller.perform_caching && app.config.action_dispatch.rack_cache
+              app.config.middleware.insert 0, ::Rack::Cache, {
+                verbose: true,
+                metastore: URI.encode("file:#{Rails.root}/tmp/dragonfly/cache/meta"), # URI encoded in case of spaces
+                entitystore: URI.encode("file:#{Rails.root}/tmp/dragonfly/cache/body")
+              }
+            end
+            app.config.middleware.insert_after ::Rack::Cache, ::Dragonfly::Middleware, :refinery_images
           else
-            app.config.middleware.use 'Dragonfly::Middleware', :refinery_images
+            app.config.middleware.use ::Dragonfly::Middleware, :refinery_images
           end
         end
       end
