@@ -8,7 +8,7 @@ module Refinery
 
       config.autoload_paths += %W( #{config.root}/lib )
 
-      initializer 'attach-refinery-images-with-dragonfly', :after => :load_config_initializers do |app|
+      initializer 'attach-refinery-images-with-dragonfly', :before => :finisher_hook do |app|
         ::Refinery::Images::Dragonfly.configure!
         ::Refinery::Images::Dragonfly.attach!(app)
       end
