@@ -7,7 +7,7 @@ module Refinery
     validates :title, :presence => true, :uniqueness => {:scope => :refinery_page_id}
     alias_attribute :content, :body
 
-    translates :body if respond_to?(:translates)
+    translates :body
 
     def to_param
       "page_part_#{title.downcase.gsub(/\W/, '_')}"
@@ -19,11 +19,11 @@ module Refinery
       normalise_text_fields
     end
 
-    self.translation_class.send :attr_accessible, :locale if self.respond_to?(:translation_class)
+    self.translation_class.send :attr_accessible, :locale
 
-  protected
+    protected
     def normalise_text_fields
-      if body.present? && body !~ %r{^<}
+      if body? && body !~ %r{^<}
         self.body = "<p>#{body.gsub("\r\n\r\n", "</p><p>").gsub("\r\n", "<br/>")}</p>"
       end
     end
