@@ -4,10 +4,9 @@ module Refinery
 
     config_accessor :pages_per_dialog, :pages_per_admin_index, :new_page_parts,
                     :marketable_urls, :approximate_ascii, :strip_non_ascii,
-                    :default_parts, :use_custom_slugs, :scope_slug_by_parent,
-                    :cache_pages_backend, :cache_pages_full, :layout_template_whitelist,
-                    :use_layout_templates, :page_title, :absolute_page_links, :types,
-                    :auto_expand_admin_tree, :show_title_in_body
+                    :cache_pages_full, :default_parts, :scope_slug_by_parent,
+                    :layout_template_whitelist, :use_layout_templates, :page_title,
+                    :absolute_page_links, :types, :auto_expand_admin_tree, :show_title_in_body
 
     self.pages_per_dialog = 14
     self.pages_per_admin_index = 20
@@ -16,11 +15,9 @@ module Refinery
     self.approximate_ascii = false
     self.strip_non_ascii = false
     self.default_parts = ["Body", "Side Body"]
-    self.use_custom_slugs = false
     self.scope_slug_by_parent = true
-    self.cache_pages_backend = false
     self.cache_pages_full = false
-    self.layout_template_whitelist = ["application"]
+    self.layout_template_whitelist = %w[application]
     class << self
       def layout_template_whitelist
         Array(config.layout_template_whitelist).map(&:to_s)
