@@ -2,30 +2,34 @@ module Refinery
   module Testing
     module ControllerMacros
       module Methods
-        def get(action, parameters = nil, session = nil, flash = nil)
-          process_refinery_action(action, parameters, session, flash, "GET")
+        def get(action, options = {})
+          process_refinery_action(action, 'GET', options)
         end
 
         # Executes a request simulating POST HTTP method and set/volley the response
-        def post(action, parameters = nil, session = nil, flash = nil)
-          process_refinery_action(action, parameters, session, flash, "POST")
+        def post(action, options = {})
+          process_refinery_action(action, 'POST', options)
         end
 
         # Executes a request simulating PUT HTTP method and set/volley the response
-        def put(action, parameters = nil, session = nil, flash = nil)
-          process_refinery_action(action, parameters, session, flash, "PUT")
+        def put(action, options = {})
+          process_refinery_action(action, 'PUT', options)
+        end
+
+        # Executes a request simulating PATCH HTTP method and set/volley the response
+        def patch(action, options = {})
+          process_refinery_action(action, 'PATCH', options)
         end
 
         # Executes a request simulating DELETE HTTP method and set/volley the response
-        def delete(action, parameters = nil, session = nil, flash = nil)
-          process_refinery_action(action, parameters, session, flash, "DELETE")
+        def delete(action, options = {})
+          process_refinery_action(action, 'DELETE', options)
         end
 
         private
 
-        def process_refinery_action(action, parameters = nil, session = nil, flash = nil, method = "GET")
-          parameters ||= {}
-          process(action, parameters.merge!(:use_route => :refinery), session, flash, method)
+        def process_refinery_action(action, http_method = 'GET', options)
+          process(action, http_method, options.merge(:use_route => :refinery))
         end
       end
     end
